+++ packages/frontend/src/hooks/useQueryResults.tsx
@@ -1,6 +1,5 @@
 import { ApiError, ApiQueryResults, MetricQuery } from 'common';
 import { useQuery } from 'react-query';
-import { useEffect } from 'react';
 import { lightdashApi } from '../api';
 import { useExplorer } from '../providers/ExplorerProvider';
 import { useApp } from '../providers/AppProvider';
@@ -39,20 +38,15 @@ export const useQueryResults = () => {
         ),
     };
     const queryKey = ['queryResults', tableId, metricQuery];
-    const query = useQuery<ApiQueryResults, ApiError>({
+    return useQuery<ApiQueryResults, ApiError>({
         queryKey,
         queryFn: () => getQueryResults(tableId || '', metricQuery),
         enabled: false, // don't run automatically
         keepPreviousData: true, // changing the query won't update results until fetch
         retry: false,
-    });
-
-    useEffect(() => {
-        if (query.error) {
-            const [first, ...rest] = query.error.error.message.split('\n');
+        onError: (error) => {
+            const [first, ...rest] = error.error.message.split('\n');
             showError({ title: first, body: rest.join('\n') });
-        }
-    }, [query.error, showError]);
-
-    return query;
+        },
+    });
 };